+++ shared/connectionmode/mode.go
