+++ shared/connectionmode/mode_test.go
